refactor: create a new unit view to hold the header and content for page
