+++ src/app/(core)/programmes/[subjectPhaseSlug]/units/[unitSlug]/lessons/page.tsx
@@ -2,9 +2,7 @@ import { Metadata } from "next";
 import { notFound } from "next/navigation";
 import { cache } from "react";
 
-import UnitHeader from "./Components/UnitHeader/UnitHeader";
 import { UnitView } from "./Components/UnitView";
-import { Breadcrumbs } from "./Components/Breadcrumbs/Breadcrumbs";
 
 import { getOpenGraphMetadata, getTwitterMetadata } from "@/app/metadata";
 import withPageErrorHandling, {
@@ -12,9 +10,6 @@ import withPageErrorHandling, {
 } from "@/hocs/withPageErrorHandling";
 import curriculumApi2023 from "@/node-lib/curriculum-api-2023";
 import { getFeatureFlagValue } from "@/utils/featureFlags";
-import { SubjectIcon } from "@/components/TeacherComponents/Header/Header";
-import { KeyStageTitleValueType } from "@/browser-lib/avo/Avo";
-import { getTeacherSubjectPhaseSlug } from "@/utils/curriculum/slugs";
 
 type LessonsPageParams = { subjectPhaseSlug: string; unitSlug: string };
 
@@ -72,72 +67,8 @@ const InnerUnitPage = async (props: AppPageProps<LessonsPageParams>) => {
 
   const { subjectPhaseSlug: programmeSlug, unitSlug } = await props.params;
   const data = await getCachedUnitData(programmeSlug, unitSlug);
-  const subjectIconName = `subject-${data.subjectSlug}` as SubjectIcon;
 
-  const subjectPhaseSlug = getTeacherSubjectPhaseSlug({
-    subjectSlug: data.subjectSlug,
-    phaseSlug: data.phaseSlug,
-    subjectParentTitle: data.parentSubject,
-    examboardSlug: data.examBoardSlug,
-    pathwaySlug: data.pathwaySlug,
-  });
-
-  return (
-    <>
-      <UnitHeader
-        heading={data.unitTitle}
-        phase={data.phaseSlug}
-        subjectPhaseSlug={subjectPhaseSlug}
-        programmeSlug={programmeSlug}
-        subjectIcon={subjectIconName}
-        nextUnit={data.nextUnit}
-        prevUnit={data.prevUnit}
-        unitDownloadFileId={
-          unitSlug.endsWith(data.unitvariantId.toString())
-            ? unitSlug
-            : `${unitSlug}-${data.unitvariantId}`
-        }
-        isGeorestrictedUnit={data.containsGeorestrictedLessons}
-        trackingProps={{
-          unitName: data.unitTitle,
-          unitSlug: data.unitSlug,
-          keyStageSlug: data.keyStageSlug,
-          keyStageTitle: data.keyStageTitle as KeyStageTitleValueType,
-          subjectSlug: data.subjectSlug,
-          subjectTitle: data.subjectTitle,
-        }}
-        headerSlot={
-          <Breadcrumbs
-            data={data}
-            subjectPhaseSlug={subjectPhaseSlug}
-            mode="unit"
-          />
-        }
-      />
-      <UnitView
-        programmeSlug={data.programmeSlug}
-        unitSlug={data.unitSlug}
-        unitTitle={data.unitTitle}
-        unitDescription={data.unitDescription}
-        subjectTitle={data.subjectTitle}
-        subjectSlug={data.subjectSlug}
-        keyStageSlug={data.keyStageSlug}
-        keyStageTitle={data.keyStageTitle}
-        lessons={data.lessons}
-        unitIndex={data.unitIndex}
-        unitCount={data.unitCount}
-        whyThisWhyNow={data.whyThisWhyNow}
-        priorKnowledgeRequirements={data.priorKnowledgeRequirements}
-        threads={data.threads}
-        phaseSlug={data.phaseSlug}
-        tierOptionToggles={data.tierOptionToggles}
-        subjectOptionToggles={data.subjectOptionToggles}
-        nextUnit={data.nextUnit}
-        prevUnit={data.prevUnit}
-        subjectCategories={data.subjectCategories}
-      />
-    </>
-  );
+  return <UnitView {...data} />;
 };
 
 const UnitPage = withPageErrorHandling(InnerUnitPage, "unit-page::app");
